+++ apps/src/tests/index.ts
@@ -150,7 +150,7 @@ export { default as Test3115 } from './Test3115';
 export { default as Test3168 } from './Test3168';
 export { default as Test3173 } from './Test3173';
 export { default as Test3239 } from './Test3239';
-export { default as TestScrollViewHorizontal } from './TestScrollViewHorizontal';
+export { default as Test3265 } from './Test3265';
 export { default as TestScreenAnimation } from './TestScreenAnimation';
 export { default as TestScreenAnimationV5 } from './TestScreenAnimationV5';
 export { default as TestHeader } from './TestHeader';
